+++ BigBang/graphlet/dashboard/thermometerlet.hpp
@@ -6,24 +6,24 @@
 #include "brushes.hxx"
 
 namespace WarGrey::SCADA {
-	private class Thermometerlet : public WarGrey::SCADA::IRangelet<float> {
+	private class Thermometerlet : public WarGrey::SCADA::IRangelet<double> {
 	public:
-		Thermometerlet(float tmin, float tmax, unsigned int step,
+		Thermometerlet(double tmin, double tmax, unsigned int step,
 			float width, float height = 0.0F, float thickness = 3.0F,
 			Microsoft::Graphics::Canvas::Brushes::ICanvasBrush^ border_color = nullptr,
 			WarGrey::SCADA::GradientStops^ colors = nullptr);
 
 		Thermometerlet(WarGrey::SCADA::FitPosition mark_position,
-			float tmin, float tmax, unsigned int step, float width, float height = 0.0F, float thickness = 3.0F,
+			double tmin, double tmax, unsigned int step, float width, float height = 0.0F, float thickness = 3.0F,
 			Microsoft::Graphics::Canvas::Brushes::ICanvasBrush^ border_color = nullptr,
 			WarGrey::SCADA::GradientStops^ colors = nullptr);
 
-		Thermometerlet(float range, unsigned int step, float width,
+		Thermometerlet(double range, unsigned int step, float width,
 			float height = 0.0F, float thickness = 3.0F,
 			Microsoft::Graphics::Canvas::Brushes::ICanvasBrush^ border_color = nullptr,
 			WarGrey::SCADA::GradientStops^ colors = nullptr);
 
-		Thermometerlet(WarGrey::SCADA::FitPosition mark_position, float range, unsigned int step,
+		Thermometerlet(WarGrey::SCADA::FitPosition mark_position, double range, unsigned int step,
 			float width, float height = 0.0F, float thickness = 3.0F,
 			Microsoft::Graphics::Canvas::Brushes::ICanvasBrush^ border_color = nullptr,
 			WarGrey::SCADA::GradientStops^ colors = nullptr);
@@ -44,10 +44,10 @@ namespace WarGrey::SCADA {
 
 	public:
 		void fill_mercury_extent(float* x = nullptr, float* y = nullptr, float* width = nullptr, float* height = nullptr);
-		void fill_mercury_extent(float percentage, float* x = nullptr, float* y = nullptr, float* width = nullptr, float* height = nullptr);
+		void fill_mercury_extent(double percentage, float* x = nullptr, float* y = nullptr, float* width = nullptr, float* height = nullptr);
 
 	protected:
-		void on_value_changed(float t) override;
+		void on_value_changed(double t) override;
 
 	private:
 		WarGrey::SCADA::GradientStops^ colors;
